[fix]ユーザーの口コミを掲載ホテルの表示をコンポーネント内でmapで回すように修正#5
